Turn tests back on
