+++ examples/default-tags-yaml/step2/Pulumi.yaml
@@ -12,4 +12,4 @@ resources:
   my-bucket:
     type: aws:s3:Bucket
 outputs:
-  bucketTags: ${my-bucket.tagsAll}
+  bucketTags: ${my-bucket.tags}
