+++ examples/default-tags-yaml/step3/Pulumi.yaml
@@ -13,4 +13,4 @@ resources:
   my-bucket:
     type: aws:s3:Bucket
 outputs:
-  bucketTags: ${my-bucket.tagsAll}
+  bucketTags: ${my-bucket.tags}
