+++ examples/default-tags-yaml/step5/Pulumi.yaml
@@ -8,4 +8,4 @@ resources:
   my-bucket:
     type: aws:s3:Bucket
 outputs:
-  bucketTags: ${my-bucket.tagsAll}
+  bucketTags: ${my-bucket.tags}
